telegram-pipeline/requirements.txt
gspread==6.1.4
google-auth==2.38.0
requests==2.32.3
colorama==0.4.6
python-dotenv==1.0.1
